Refactor : Security Config 충돌 수정
