+++ Server/src/main/java/ForMZ/Server/global/config/SecurityConfig.java
@@ -5,7 +5,10 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.crypto.factory.PasswordEncoderFactories;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
 
 @Configuration
 @EnableWebSecurity
@@ -17,6 +20,23 @@ public class SecurityConfig {
                 .csrf(AbstractHttpConfigurer::disable)
                 .formLogin(AbstractHttpConfigurer::disable);
 
+        http
+                .authorizeHttpRequests(auth -> auth
+                        .requestMatchers(
+                                new AntPathRequestMatcher("/login"),
+                                new AntPathRequestMatcher("/email")
+                        ).permitAll()
+                        .requestMatchers(
+                                new AntPathRequestMatcher("/my-page")
+                        ).hasRole("USER")
+                        .anyRequest().authenticated());
+
         return http.build();
+    }
+
+    @Bean
+    public PasswordEncoder passwordEncoder(){
+        // DelegatingPasswordEncoder : Spring Security의 권장 알고리즘을 사용하여 비밀번호를 인코딩
+        return PasswordEncoderFactories.createDelegatingPasswordEncoder();
     }
 }
